oooast-tools-20 DbEntityClass refactored
